group reward code by type
